requirements.txt
streamlit
langchain-groq
langchain
langchain-community
sqlalchemy
mysql-connector-python
python-dotenv
